requirements.txt
pytest==7.2.0
requests==2.31.0
python-dotenv=1.0.1
jsonschema=4.23.0